SC-2323 Refactoring User bulk upload Module
