+++ service/src/main/java/org/sunbird/service/user/impl/UserServiceImpl.java
@@ -25,7 +25,6 @@ import org.sunbird.keys.JsonKey;
 import org.sunbird.logging.LoggerUtil;
 import org.sunbird.model.adminutil.AdminUtilRequestData;
 import org.sunbird.model.user.User;
-import org.sunbird.operations.ActorOperations;
 import org.sunbird.request.Request;
 import org.sunbird.request.RequestContext;
 import org.sunbird.response.Response;
