+++ service/src/test/java/org/sunbird/actor/user/UserSelfDeclarationManagementActorTest.java
@@ -53,7 +53,7 @@ import scala.concurrent.Promise;
   DataCacheHandler.class,
   org.sunbird.datasecurity.impl.ServiceFactory.class,
   EmailTemplateDaoImpl.class,
-  Util.class,
+  UserUtility.class,
   EsClientFactory.class,
   ElasticSearchRestHighImpl.class,
   UserUtil.class,
@@ -113,7 +113,7 @@ public class UserSelfDeclarationManagementActorTest {
             Mockito.anyString(), Mockito.anyString(), Mockito.anyMap(), Mockito.any()))
         .thenReturn(response);
 
-    PowerMockito.mockStatic(Util.class);
+    PowerMockito.mockStatic(UserUtility.class);
     when(UserUtility.encryptData(Mockito.anyString())).thenReturn("userExtId");
 
     PowerMockito.mockStatic(UserUtil.class);
