+++ service/src/test/java/org/sunbird/service/organisation/OrgServiceImplTest.java
@@ -207,4 +207,27 @@ public class OrgServiceImplTest {
     res.put(JsonKey.RESPONSE, JsonKey.SUCCESS);
     return res;
   }
+
+  @Test
+  public void testRegisterChannel() {
+    OrgService orgService = OrgServiceImpl.getInstance();
+    Map<String, Object> map = new HashMap<>();
+    map.put(JsonKey.CHANNEL, "ch");
+    map.put(JsonKey.DESCRIPTION, "desc");
+    map.put(JsonKey.ID, "id");
+    Boolean bool = orgService.registerChannel(map, new RequestContext());
+    Assert.assertNotNull(bool);
+  }
+
+  @Test
+  public void testUpdateChannel() {
+    OrgService orgService = OrgServiceImpl.getInstance();
+    Map<String, Object> map = new HashMap<>();
+    map.put(JsonKey.CHANNEL, "ch");
+    map.put(JsonKey.DESCRIPTION, "desc");
+    map.put(JsonKey.ID, "id");
+    Boolean bool = orgService.updateChannel(map, new RequestContext());
+    Assert.assertNotNull(bool);
+  }
+
 }
